HOMME: Bringing in the work from HOMMEXX as a new target in the multimodel framework
The preqx_kokkos target is tested against the preqx one, for a few configurations.

We also bring in the hommexx testing framework, consisting of different levels
of testing. This allows to run tests at specific resolutions, or even
just unit tests. However, all the preqx_kokkos tests (except for unit tests)
are not yet enabled.